外來鍵欄位屬性修改

+++ database/migrations/2021_03_17_070240_create_favorites_table.php
@@ -15,8 +15,8 @@ class CreateFavoritesTable extends Migration
     {
         Schema::create('favorites', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('user_id');
-            $table->integer('trail_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('trail_id');
             $table->timestamps();
         });
     }
