+++ database/migrations/2021_03_31_130858_create_user_like_comments_table.php
@@ -15,8 +15,8 @@ class CreateUserLikeCommentsTable extends Migration
     {
         Schema::create('user_like_comments', function (Blueprint $table) {
             $table->id();
-            $table->integer('comment_id');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('comment_id');
+            $table->unsignedBigInteger('user_id');
             $table->integer('status');
             $table->timestamps();
         });
